lattigo: emit tensor.splat
I think the right way to do this is to reuse the std library
Repeat function (https://pkg.go.dev/slices#Repeat).

The complexity here is that the import cannot be unused, so I changed
the emitter to dynamically track a list of imports needed by the emitted
program, and then only emit the prelude at the end, which requires
passing in a separate ostream at the start and combining it with the
prelude at the end.